mpscfifo is now functional for single thread.
CAREFUL: None of this has not been tested using multiple threads!

Add extra pointer to LinkNode as my initial attempt at using inheritance
didn't work.

There were three bugs first was rmvNode was returning the next node which
is active, not the unused node which is pointer to by head.

Secondly in msgarea.getLinkNode I needed the initialize the link node
if we are reusing a previously allocated node.

Third delMpmcStack wasn't working we needed to pop and delete the nodes
in msgarena and also to not leak the initial MpscFifo link we are returning
mq.head when we delMpscFifo.

Moved LinkNode procs to linknode.nim

Added additional tests.