Fixed the delMpscFifo bug.
The problem was when rmvNode returns the stub node we needed to initialize
result.next to nil so when its reused its nil. Otherwise the queue looks
not empty if that LinkNode is added to an empty fifo. I also added
initLinkNode so that callers may initialize an existing LinkNode although
that's obviously optional now.

For the time being I'm having LinkNode assume extra is a Msg.