Added lnlooper.
lnlooper is a slight modification to msglooper instead instead of
ProcessMsg taking a MsgPtr it takes a LinkNodePtr.

This test was to see what performance difference this might make, in
particular if it would help the thread starvation problem I see as the
number of threads increases above 24. At 48 I see about a 2:1 difference
between the minimum mean and the maximum mean.

Overall the performance was fairly close with the LinkNodePtr version
definitely a 5% or so faster above 24 threads but less differences at the
for instance 1 thread. I actually saw bm3 was slightly faster with a mean
of 17.54 where as bm4 was 18.01 and other times bm4 was faster with a
mean of 17.50 and bm3 at 18.52 so in the noise.

So the LinkNodePtr version was better the there is still significant
thread starvation, so apparently this is more due to the OS and less
to do with the application algorithms.

Note: these tests were run without -d:release and I did see bm3 fail once
at line 93 in mpmcstack.nim.