Remove LinkNode and have Msg have next.
This simplifies things but is a just a little bit slower in the simplest
case because of the additional work to move rspq and cmd when we convert
the Msg into a stub and maybe because of alignment. Its a little slower
than I thought in the optimal case of bm1 non-blocking the mean was
29cy now it's 40cy, but the min is the same.

old:
  # mpscfifo.test add/rmvNode non-blocking: ts={min=10cy mean=29cy minC=7763 n=5480653}
new:
  # mpscfifo.bm add/rmv non-blocking: ts={min=10cy mean=40cy minC=9 n=5195976}


I also added tests where we do a get/retMsg each time and its definitely
slower but probably not nearly as bad if we used the old blocking code
for get/ret. I added this to the bm3 tests and the get/retMsg is lost in
the overhead of invoking the OS for the cross thread communication.

A further improvement will be to use a wait-free algorithm instead of the
lock-free algorithm for get/retMsg.