Remove MsgArena.lock
The lock is not providing any effective mutual exclusion since I changed to
using MpmcStack for saving the messages.

But doing this causes mpscfifo to crash when deleting the fifo after using
get/retMsg:

 run: ./mpscfifo
 [OK] test new queue is empty
 [OK] test new queue is empty twice
 [OK] test add, rmv blocking
 Traceback (most recent call last)
 mpscfifo.nim(287)        mpscfifo
 mpscfifo.nim(96)         newMpscFifo
 mpscfifo.nim(64)         newMpscFifo
 gc.nim(269)              unsureAsgnRef
 gc.nim(167)              decRef
 SIGSEGV: Illegal storage access. (Attempt to read from nil?)

I'm raising an issue and the quick fix is either to not do this or not
use the name, for now I'm going to not use the name.