fix allocation/dealloction.
Add fifoutils which has initializer, ptrToStr, echoBytes and deleted
the individual files.

The initializer is the most important and is used by objects allocated
on the shared heap. This slowed things down and for critical performance
area there will eventually be optimizations, but this is good for now as
it fixes at least one weird bug in mpscfifo which causes a SIGSEGV when
mq.name is initialized to name. Previously this was fixed by using
allocShared0.