Partial hierarchical support.
Added a states Table which with a key of a state and a parentStateInfo.
Currently nothing is done with the hierarchy but you can add states and
set their parent.

Added some tests for one, two states plus a default with two child states
also this last test make no attempt to pass "unhandled" messages which
should be handled by the parent. Something for the future.

Add StateMachine.java the hierarchical state machine I created for Android.
This is there as an example.