Added delComponent.
This is partially functional the main thing missing at the moment is
to either make add/delComponent fully synchronous. Actually a nicer
option would be to allow asynchronous or using Future's. The caller
then has the best of both worlds.

Adding the MsgProcessorState feels pretty nice, but there seems to be
at least one bug in the compiler. I couldn't have a forward decleartion
that the looper called because the compiler complained looper wasn't
gcsafe. But if I had the routine defined before looper it worked.
A I also got a gcsafe issue whein I tried to tests mp.newComponent
or mp.delComponent in an if statement. I think both should work, but I
need to have a simpler test case.