Add support for not handling a state.
When a state is not handle its parent state is asked to handle it. So
StateResult has been added and each TypeState returns Handled or
NotHandled. Also, callers which don't care about the result discard it.

The dispatcher has been modified to allow parents to handle NotHandled
messages.

invokeExitProcs was fixed a bug where the stateStack wasn't popped
correctly.

Added more DBG and now test-triangle-hsm tests a NotHandled case.